+++ src/config.ts
@@ -27,6 +27,10 @@ export function get(key: "playgrounds.templateGalleries"): string[];
 export function get(key: "comments.showThread"): string;
 export function get(key: any) {
   const extensionConfig = vscode.workspace.getConfiguration(CONFIG_SECTION);
-
   return extensionConfig.get(key);
+}
+
+export async function set(key: string, value: any) {
+  const extensionConfig = vscode.workspace.getConfiguration(CONFIG_SECTION);
+  return extensionConfig.update(key, value, true);
 }